add the feeCall to estimateGas if we're using a paymaster or the network is predefined
